+++ libgraph/vector_masked_max_index.hh
@@ -52,7 +52,7 @@ namespace honei
                 throw VectorSizeDoesNotMatch(mask.size(), vector.size());
 
 
-            int i(0);
+            unsigned long i(0);
             for( ; i < mask.size() ; ++i)
             {
                 if (mask[i] == 1)
@@ -91,7 +91,7 @@ namespace honei
             if (vector.size() != mask.size())
                 throw VectorSizeDoesNotMatch(mask.size(), vector.size());
 
-            int i(0);
+            unsigned long i(0);
             for( ; i < mask.size() ; ++i)
             {
                 if (mask[i] == 1)
